+++ gammapy/utils/docs.py
@@ -25,8 +25,11 @@ from docutils import nodes
 from docutils.parsers.rst import roles
 from docutils.parsers.rst.directives import register_directive
 from docutils.parsers.rst.directives.images import Image
+from docutils.parsers.rst.directives.misc import Include
+from docutils.parsers.rst.directives.body import CodeBlock
 from nbformat.v4 import new_markdown_cell
 from sphinx.util import logging
+from gammapy.scripts import Config
 from gammapy import __version__
 
 try:
@@ -36,6 +39,31 @@ except KeyError:
     HAS_GP_DATA = False
 
 log = logging.getLogger("__name__")
+
+
+class HowtoHLI(Include):
+    """Directive to insert how-to for high-level interface"""
+
+    def run(self):
+        raw = ""
+        section = self.arguments[0]
+        doc = Config._get_doc_sections()
+        for keyword in doc.keys():
+            if section == "" or section == keyword:
+                raw += doc[keyword]
+        include_lines = raw.splitlines()
+        codeblock = CodeBlock(
+            self.name,
+            [],
+            self.options,
+            include_lines,  # content
+            self.lineno,
+            self.content_offset,
+            self.block_text,
+            self.state,
+            self.state_machine,
+        )
+        return codeblock.run()
 
 
 class DocsImage(Image):
@@ -111,6 +139,7 @@ def gammapy_sphinx_ext_activate():
     # Register our directives and roles with Sphinx
     register_directive("gp-image", DocsImage)
     roles.register_local_role("gp-notebook", LinkNotebook)
+    register_directive("gp-howto-hli", HowtoHLI)
 
 
 def parse_notebooks(folder, url_docs):