+++ gammapy/modeling/iminuit.py
@@ -22,7 +22,7 @@ def optimize_iminuit(parameters, function, **kwargs):
 
     Parameters
     ----------
-    parameters : `~gammapy.utils.modeling.Parameters`
+    parameters : `~gammapy.modeling.Parameters`
         Parameters with starting values
     function : callable
         Likelihood function