+++ gammapy/modeling/models/spectral.py
@@ -14,13 +14,7 @@ from gammapy.utils.scripts import make_path
 
 
 class SpectralModel(Model):
-    """Spectral model base class.
-
-    Derived classes should store their parameters as
-    `~gammapy.utils.modeling.Parameters`
-    See for example return pardict of
-    `~gammapy.modeling.models.PowerLawSpectralModel`.
-    """
+    """Spectral model base class."""
 
     def __call__(self, energy):
         kwargs = {}
@@ -866,7 +860,7 @@ class ExpCutoffPowerLawSpectralModel(SpectralModel):
 
     @property
     def e_peak(self):
-        r"""Spectral energy distribution peak energy (`~astropy.utils.Quantity`).
+        r"""Spectral energy distribution peak energy (`~astropy.units.Quantity`).
 
         This is the peak in E^2 x dN/dE and is given by:
 
@@ -1616,7 +1610,7 @@ class NaimaSpectralModel(SpectralModel):
     likelihood spectral fit of the model's parameters directly on observations, as opposed to the MCMC
     `fit to flux points <https://naima.readthedocs.io/en/latest/mcmc.html>`_ featured in
     Naima. All the parameters defining the parent population of charged particles are stored as
-    `~gammapy.utils.modeling.Parameter` and left free by default. In case that the radiative model is `
+    `~gammapy.modeling.Parameter` and left free by default. In case that the radiative model is `
     ~naima.radiative.Synchrotron`, the magnetic field strength may also be fitted. Parameters can be
     freezed/unfreezed before the fit, and maximum/minimum values can be set to limit the parameters space to
     the physically interesting region.