+++ gammapy/utils/__init__.py
@@ -7,6 +7,6 @@ the `gammapy.utils` namespace is empty.
 
 Examples::
 
-    from gammapy.utils.modeling import Parameter
+    from gammapy.utils.interpolation import ScaledRegularGridInterpolator
     from gammapy.utils.nddata import NDDataArray
 """